requirements.txt
altair==4.2.2
applicationinsights==0.11.10
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.1
azure-functions==1.19.0
azure-identity==1.16.0
azure-mgmt-compute==31.0.0
azure-mgmt-core==1.4.0
azure-mgmt-costmanagement==4.0.1
azure-mgmt-monitor==6.0.2
azure-mgmt-network==25.3.0
azure-mgmt-resource==23.1.1
azure-mgmt-sql==3.0.1
azure-mgmt-storage==21.1.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
entrypoints==0.4
fonttools==4.51.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
isodate==0.6.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
msal==1.28.0
msal-extensions==1.1.0
msrest==0.7.1
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
patsy==0.5.6
pillow==10.3.0
portalocker==2.8.2
prettytable==3.10.0
protobuf==4.25.3
pyarrow==16.1.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
smmap==5.0.1
statsmodels==0.14.2
streamlit==1.35.0
streamlit-aggrid==1.0.5
tabulate==0.9.0
tenacity==8.3.0
termcolor==2.4.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.1
wcwidth==0.2.13
